Add History protocol
